# public/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Next Swagger API Aqua-Track",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api"
        },
        {
            "url": "https://aqua-track-olive.vercel.app/api"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "csrfTokenAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "next-auth.csrf-token",
                "description": "After the first login, the session token is stored in cookies, and there is no need to retrieve the csrf token again during subsequent request actions. If you clear your cookies, you'll need to request the csrf token again. In this field, it is not necessary to enter a value for the header, because it still needs to be passed in the body of the authorization request through the selected type (google or credentials) to the csrfToken field. You can also log in to the site in the same browser as swagger-docs and not enter tokens manually. They will be automatically extracted from cookies."
            }
        },
        "schemas": {
            "Water": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "The unique identifier for the water entry (ObjectId)."
                    },
                    "userId": {
                        "type": "string",
                        "description": "The unique identifier for the user (ObjectId)."
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "description": "The date when the water entry was recorded (YYYY-MM-DD format)."
                    },
                    "time": {
                        "type": "string",
                        "description": "The time when the water was consumed."
                    },
                    "volume": {
                        "type": "integer",
                        "description": "The volume of water consumed (in milliliters)."
                    }
                },
                "required": ["date", "time", "volume"],
                "description": "Represents a water consumption entry for a user."
            },
...
        },
...
    },
    "paths": {
...
        "/water/monthly": {
            "get": {
                "tags": ["Water"],
                "summary": "Get monthly water entry",
                "security": [
                    {
                        "csrfTokenAuth": []
                    }
                ],
                "description": "Retrieves the water entry for a specific date using query parameter.",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "description": "Date for which to retrieve the water entry, formatted as YYYY-MM-DD.",
                        "schema": {
                            "type": "string",
                            "example": "2024-09-25"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monthly water entry retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Water"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    }
}